requirements.txt
discord.py>=2.3
google-genai>=1.0
python-dotenv>=1.0
apscheduler>=3.10
aiosqlite>=0.19
